muevo template home_profesionales al modulo de profesionales. Fixed #97
