--- ggg/templates/home_profesional.html
@@ -1,37 +1,0 @@
-{% extends "base_admin.html" %}
-
-{% block content %}
-{% if not request.user.is_authenticated %}
-<a class="btn btn-link btn-primary" href="{% url 'account_login' %}"> Acceso </a>
-{% endif %}
-
-    <table class="table">
-        <thead>
-        <tr>
-            <th scope="col">Paciente</th>
-            <th scope="col">Inicio</th>
-            <th scope="col">Fin</th>
-            <th scope="col">Servicio</th>
-            <th scope="col">Acciones</th>
-        </tr>
-        </thead>
-        <tbody>
-            {% for turno in turnos %}
-                <tr>
-                    <th scope="row">{{ turno.paciente.nombres }} {{ turno.paciente.apellidos }}</th>
-                    <td>{{ turno.inicio }}</td>
-                    <td>{{ turno.fin }}</td>
-                    <td>{{ turno.servicio.especialidad.nombre }}</td>
-                    <th scope="row">
-                        <a role="button" class="btn btn-warning btn-sm active" href="{% url 'profesionales.crear.consulta' dni=turno.paciente.numero_documento prof=turno.profesional.id %}">ingresar</a>
-                    </th>
-                </tr>
-                {% empty %}
-                <td>Sin datos</td><td></td><td></td>
-            {% endfor %}
-        </tbody>
-    </table>
-
-{% include 'admin_paginate.html' %}
-
-{% endblock %}
